+++ CHANGELOG.md
@@ -54,6 +54,7 @@ Compatibility:
 * Sort by default for `Dir.{glob,[]}` and add `sort:` keyword argument (#2523, @Strech).
 * Implement `rb_str_locktmp` and `rb_str_unlocktmp` (#2524, @bjfish).
 * Update `Kernel#instance_variables` to return insertion order (@bjfish).
+* Fixed `rb_path2class()` to not error for a module (#2511, @eregon).
 
 Performance: